Remove colorizing from logging
I don't need the log to be specially colored, just do it in a simple way. This fixes a Rails 7.1 deprecation:

```
DEPRECATION WARNING: Bolding log text with a positional boolean is deprecated and will be removed in Rails 7.2. Use an option hash instead (eg. `color("my text", :red, bold: true)`)
```